Smart
Hi, I am Bashir Ebrahimi from Iran. I came up with an idea for blockchain and smart contracts, this idea can be very effective and can be easily implemented for many jobs in the world and anywhere in the world. I have studied and experienced blockchains and cryptocurrencies for years to come up with this idea, now this idea is my dream and can even increase the value of your company. In return for implementing this important idea and project, I want to partner with each other and enjoy the benefits of a partnership with a lifetime contract, because implementing this project will always be profitable and efficient. And that for each stadium or team in which the advertisement is done, a token or a special Cryptocurrency is produced and paid to the companies that provide the advertisement by proving the shares, and thus the income, quality and number of advertisements. Increases and also the desired currency becomes valuable. With the implementation of this project, billboards will no longer be limited to one ad, and each billboard can be assigned to one ad from one company, thus giving more ads to the places where the ads are displayed, as well as increasing their income and thus fame. And the speed of introducing companies increases and their income also increases. This idea is very valuable and can save even sports clubs from bankruptcy in this situation of coronavirus and other cases and increase their income and even boost many businesses and companies and other businesses.
I wanted to name this idea smart advertising, because smart advertising not only on digital billboards in stadiums and other places, but everywhere and everywhere in the world, including social networks. And media will also be playable and executable. Smart advertising can be both revenue-generating. These ideas or advertisements can be in the form of a token or a special currency to be traded with.

By implementing this idea, it is possible to expand advertisements everywhere and everywhere, except for digital billboards on sites and everywhere, automatically and online, and this will help to generate more revenue and introduce advertisers, as well as to all displays. Advertisers also generate high revenue, which means increasing the value of blockchain and the value of businesses and their revenue and unlimited economies around the world.
And that for each stadium or team in which the advertisement is done, a token or a special Cryptocurrency is produced and paid to the companies that provide the advertisement by proving the shares, and thus the income, quality and number of advertisements. Increases and also the desired currency becomes valuable. With the implementation of this project, billboards will no longer be limited to one ad, and each billboard can be assigned to one ad from one company, thus giving more ads to the places where the ads are displayed, as well as increasing their income and thus fame. And the speed of introducing companies increases and their income also increases.